feat: add all standard html elements and attributes
Standard HTML elements as well as standard attributes are supported. Deprecated elements and attributes are not supported. Experimental elements and attributes should be added progressively.

HTML elements include a MDN description as docstring, allowing to get a quick overview of the element on hover when using a proper IDE like VSCODE.

HTML attributes support autocompletion for attributes names as well as attribute values.

Attribute values are automatically escaped and URL attributes are URL-encoded.